fix(sql): promote attempts+1 to bigint to survive an i32::MAX row
The orphan sweep (`reenqueue_orphaned_blocking`) and the undecodable-task
release (`fail_undecodable_task`) both compute `attempts + 1` on the int
`attempts` column. A row already sitting at `i32::MAX` — which normal
flow never produces (attempts is always LEAST-bounded by `max_attempts`)
but a corrupt or hand-written row can — makes that `+ 1` overflow, and
PostgreSQL aborts the whole statement with "integer out of range". A
single such row would then fail every sweep for its queue.

Promote the arithmetic to `attempts::bigint + 1` in both statements. The
`>=` comparisons no longer overflow, and `LEAST(attempts::bigint + 1,
max_attempts)` re-bounds the assigned value back into int range.

Add an integration spec that plants a `Running` row at `attempts =
i32::MAX` and asserts the sweep completes (killing the exhausted row)
instead of erroring; the SQL mirrors are updated in lock-step.

+++ src/queries/fetch.rs
@@ -146,12 +146,16 @@ pub(crate) fn fail_undecodable_task(
         // `last_result`, so readers observe one format for every failure path.
         let result = serde_json::json!({ "Err": crate::ack::truncate_error_payload(error) });
         sql_query(
+            // `attempts::bigint` promotes the arithmetic so a corrupt row at
+            // i32::MAX cannot overflow the `+ 1` (which PostgreSQL rejects as
+            // "integer out of range"); LEAST re-bounds the result to
+            // `max_attempts`, so it fits the int column again.
             "UPDATE apalis.jobs
              SET status = CASE
-                     WHEN attempts + 1 >= max_attempts THEN 'Killed'
+                     WHEN attempts::bigint + 1 >= max_attempts THEN 'Killed'
                      ELSE 'Failed'
                  END,
-                 attempts = LEAST(attempts + 1, max_attempts),
+                 attempts = LEAST(attempts::bigint + 1, max_attempts),
                  done_at = now(),
                  last_result = $3
              WHERE id = $1
